This is the same change made in #13040, plus subsequent improvements, applied to agent-mode code.

Brought in some tests too, adapted for the different API.