@bboreham 2023-12-30
I added a benchmark:

```
name            old time/op    new time/op    delta
CreateSeries-8     593ns ± 5%     526ns ± 5%  -11.20%  (p=0.002 n=6+6)

name            old alloc/op   new alloc/op   delta
CreateSeries-8      254B ± 1%      177B ± 6%  -30.19%  (p=0.004 n=5+6)

name            old allocs/op  new allocs/op  delta
CreateSeries-8      2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.002 n=6+6)
```